requirements.txt
asgiref==3.7.2
certifi==2023.7.22
channels==4.0.0
charset-normalizer==3.2.0
colorama==0.4.6
customtkinter==0.3
darkdetect==0.8.0
distlib==0.3.8
daphne==4.1.0
Django==5.0.1
django-admin==2.0.2
django-cors-headers==3.8.0
django-excel-response2==3.0.6
django-prometheus==2.3.1
django-six==1.0.5
djangorestframework==3.14.0
excel-base==1.0.4
filelock==3.13.1
idna==3.4
isoweek==1.3.3
numpy==1.25.2
opencv-python==4.8.0.76
Pillow==10.0.0
pipenv==2023.11.15
platformdirs==4.1.0
pycryptodome==3.18.0
psycopg==3.1.8
psycopg2==2.9.1
pyJWT==2.3.0
pygame==2.1.2
pyotp==2.6.0
pypi==2.1
python-dateutil==2.8.2
qrcode==7.3
requests==2.31.0
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.0.4
virtualenv==20.25.0
xlwt==1.3.0
